+++ CHANGES.rst
@@ -63,6 +63,7 @@ Bugfixes
   been deleted in the meantime (:pr:`5386`)
 - Fix error when editing a room's nonbookable periods (:pr:`5390`)
 - Fix incorrect access check when directly accessing a registration form (:pr:`5406`)
+- Fix error in rate limiter when using Redis with a UNIX socket connection (:issue:`5391`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^